fix: rename ViewPage to SceneViewPage
