+++ src/components/SceneViewPage/SceneViewMenu/SceneViewMenu.container.ts
@@ -1,8 +1,8 @@
 import { connect } from 'react-redux'
 import { RootState } from 'modules/common/types'
-import { MapStateProps, MapDispatch, MapDispatchProps } from './ViewMenu.types'
-import ViewMenu from './ViewMenu'
 import { isLoggedIn } from 'modules/auth/selectors'
+import { MapStateProps, MapDispatch, MapDispatchProps } from './SceneViewMenu.types'
+import SceneViewMenu from './SceneViewMenu'
 
 const mapState = (state: RootState): MapStateProps => ({
   isLoggedIn: isLoggedIn(state)
@@ -12,4 +12,4 @@ const mapDispatch = (_dispatch: MapDispatch): MapDispatchProps => ({
   onTryItOut: () => null
 })
 
-export default connect(mapState, mapDispatch)(ViewMenu)
+export default connect(mapState, mapDispatch)(SceneViewMenu)
