+++ src/components/SceneViewPage/SceneViewMenu/SceneViewMenu.css
@@ -1,8 +1,8 @@
-.ViewPageMenu .made-with {
+.SceneViewMenu .made-with {
   margin: 1em;
 }
 
-.ViewPageMenu .ui.button {
+.SceneViewMenu .ui.button {
   font-size: 12px;
   line-height: 18px;
   padding: 9px 14px;
