+++ src/components/SceneViewPage/SceneViewMenu/SceneViewMenu.tsx
@@ -3,11 +3,11 @@ import { Button } from 'decentraland-ui'
 
 import { t } from 'decentraland-dapps/dist/modules/translation/utils'
 import UserMenu from 'components/UserMenu'
-import { Props, State } from './ViewMenu.types'
+import { Props, State } from './SceneViewMenu.types'
 
-import "./ViewMenu.css"
+import "./SceneViewMenu.css"
 
-export default class ViewMenu extends React.PureComponent<Props, State> {
+export default class SceneViewMenu extends React.PureComponent<Props, State> {
 
   handleTryItOut = () => {
     const { onTryItOut } = this.props
@@ -23,7 +23,7 @@ export default class ViewMenu extends React.PureComponent<Props, State> {
       return <UserMenu />
     }
 
-    return <div className="ViewPageMenu">
+    return <div className="SceneViewMenu">
       <span className="made-with">{t("public_page.made_with")}</span>
       <Button primary>{t("public_page.try_it_out")}</Button>
     </div>
