+++ src/components/SceneViewPage/SceneViewMenu/index.ts
@@ -1,0 +1,3 @@
+import SceneViewMenu from "./SceneViewMenu.container"
+
+export default SceneViewMenu
