+++ src/components/SceneViewPage/index.ts
@@ -1,0 +1,3 @@
+import SceneViewPage from './SceneViewPage.container'
+
+export default SceneViewPage
