--- src/components/ViewPage/ViewMenu/index.ts
@@ -1,3 +1,0 @@
-import ViewMenu from "./ViewMenu.container"
-
-export default ViewMenu
