--- src/components/ViewPage/index.ts
@@ -1,3 +1,0 @@
-import ViewPage from './ViewPage.container'
-
-export default ViewPage
